+++ CHANGELOG.md
@@ -2,9 +2,9 @@ CHANGE LOG
 ==========
 
 
-## V2.0 Beta 1 (Upcoming - July 2014)
+## V2.0 RC 1 (16/07/2014)
 
-* Upgrade to Laravel ~4.1
+* Upgrade to Laravel 4.1+
 * Moved to parsedown extra
 * Refactoring
 * Added di aliases